many changes back+front

+++ E-credit server/DemandeCredit/src/main/java/com/spring/DemandeCredit/Controllers/DemandeCreditController.java
@@ -25,7 +25,7 @@ public class DemandeCreditController {
     }
 
     @GetMapping("/user/{id}")
-    public ResponseEntity<List<DemandeCredit>> findAllDemandesCredit(@PathVariable("id") Integer userId) {
+    public ResponseEntity<List<DemandeCredit>> findAllDemandesCredit(@PathVariable("id") Long userId) {
         return ResponseEntity.ok(demandeCreditService.findAllDemandeCreditByUser(userId));
     }
 }
