+++ E-credit server/DemandeCredit/src/main/java/com/spring/DemandeCredit/Repositories/DemandeCreditRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DemandeCreditRepository extends JpaRepository<DemandeCredit,Integer> {
-    List<DemandeCredit> findAllByUserId(Integer userId);
+    public interface DemandeCreditRepository extends JpaRepository<DemandeCredit,Long> {
+    List<DemandeCredit> findAllByUserId(Long userId);
 }
