+++ E-credit server/DemandeCredit/src/main/java/com/spring/DemandeCredit/Services/DemandeCreditService.java
@@ -23,7 +23,7 @@ public class DemandeCreditService {
         return demandeCreditRepository.save(demandeCredit);
     }
 
-    public List<DemandeCredit> findAllDemandeCreditByUser(Integer userId) {
+    public List<DemandeCredit> findAllDemandeCreditByUser(Long userId) {
         return demandeCreditRepository.findAllByUserId(userId);
     }
 }
