+++ E-credit server/notification/src/main/java/com/spring/notification/NotifactionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface NotifactionRepository extends JpaRepository<Notification, Integer> {
-    List<Notification> findAllByUserId(Integer userId);
+public interface NotifactionRepository extends JpaRepository<Notification, Long> {
+    List<Notification> findAllByUserId(Long userId);
 }
